Added Category insert and delete
-Organized styles to be more managable and consistent
-Made modal forms exit when clicked out
-Replaced deletion alert boxes with a hidden message
-Changed window redirect in Account.js to an absolute location
-Added ability to add and delete categories
-Removed unnecessary reloading of the product grid
-made navbar contents and filter form responsive

+++ BackEnd/src/main/java/com/ministore/service/CategoryService.java
@@ -2,8 +2,10 @@ package com.ministore.service;
 
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.data.domain.PageRequest;
 import org.springframework.stereotype.Service;
 
 import com.ministore.model.Category;
@@ -19,12 +21,16 @@ public class CategoryService {
 		cr.save(c);
 	}
 	
-	public Category getByName(String name) {
-		return cr.findById(name).get();
+	public boolean has(String name) {
+		return  !cr.findById(name).isEmpty();
 	}
 	
-	public void delete(String name) {
-		cr.deleteById(name);
+	public boolean delete(String name) {
+		if (cr.countProducts(name) == 0) {
+			cr.deleteById(name);
+			return true;
+		}
+		return false;
 	}
 	
 	public List<Category> getAll() {
